PBCS/requirements.txt
ecdsa

pycryptodome

boto3